+++ src/main/java/ir/ui/golestan/exception/CourseNotFoundException.java
@@ -1,8 +1,8 @@
-package ir.ui.golestan.mamadreza;
+package ir.ui.golestan.exception;
 
 public class CourseNotFoundException extends RuntimeException{
 
-    CourseNotFoundException(int id) {
+    public CourseNotFoundException(int id) {
         super("Could not find course " + id);
     }
 }
